Extract generic CreateOptionsPanel so rules and action policies share one config-driven empty state
Moves the action-panel rendering, disabled/tooltip logic, and section-divider
into components/create_options/. Both RuleCreateOptionsPanel (list layout) and
ActionPolicyCreateOptionsPanel now delegate to the shared panel, eliminating
~150 lines of duplicated styles and component code.
