+++ x-pack/platform/plugins/shared/alerting_v2/public/components/action_policy/create_options/action_policy_create_options_panel.tsx
@@ -6,100 +6,15 @@
  */
 
 import React from 'react';
-import {
-  EuiFlexGroup,
-  EuiFlexItem,
-  EuiIcon,
-  EuiImage,
-  EuiPageTemplate,
-  EuiPanel,
-  EuiText,
-  EuiToolTip,
-  euiMinBreakpoint,
-  type UseEuiTheme,
-} from '@elastic/eui';
-import { css } from '@emotion/react';
-import { useMemoCss } from '@kbn/css-utils/public/use_memo_css';
+import { EuiImage } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n-react';
+import type { CreateOptionItem } from '../../create_options';
+import { CreateOptionsPanel } from '../../create_options';
 import type { AgentBuilderSkillsRequirements } from '../../../hooks/use_are_agent_builder_skills_available';
 import rulesListEmptyIllustration from '../../../assets/illustration-results-128.svg';
 
-export interface ActionPolicyCreateOption {
-  id: string;
-  iconType: string;
-  title: string;
-  description: string;
-  onClick: () => void;
-  /** When `true`, the option is rendered disabled and its click is a no-op. */
-  disabled?: boolean;
-  /** When set, this string is shown as a hover/focus tooltip, independent of `disabled`. */
-  tooltipText?: string;
-  'data-test-subj'?: string;
-}
-
-interface ActionPolicyCreateOptionsPanelProps {
-  options: ActionPolicyCreateOption[];
-}
-
-/** Fits the option descriptions on one line. */
-const LIST_EMPTY_STATE_MAX_INLINE_SIZE = '44em';
-
-const listEmptyStateStyles = {
-  parent: css({
-    display: 'flex',
-    flexGrow: 1,
-    height: '100%',
-    width: '100%',
-  }),
-  template: css({
-    backgroundColor: 'inherit',
-    marginInline: 'auto',
-    maxInlineSize: LIST_EMPTY_STATE_MAX_INLINE_SIZE,
-    width: '100%',
-  }),
-  widgetContainer: (euiThemeContext: UseEuiTheme) => {
-    const { euiTheme } = euiThemeContext;
-    return css({
-      padding: euiTheme.size.xl,
-      borderRadius: euiTheme.border.radius.medium,
-      width: '100%',
-      maxInlineSize: LIST_EMPTY_STATE_MAX_INLINE_SIZE,
-      boxSizing: 'border-box',
-      [euiMinBreakpoint(euiThemeContext, 'm')]: {
-        width: LIST_EMPTY_STATE_MAX_INLINE_SIZE,
-      },
-      '.euiEmptyPrompt__icon': {
-        marginBottom: euiTheme.size.l,
-        paddingRight: euiTheme.size.s,
-      },
-      '.euiEmptyPrompt__content, .euiEmptyPrompt__actions': {
-        maxInlineSize: '100%',
-        width: '100%',
-      },
-    });
-  },
-  actionsWrapper: css({
-    width: '100%',
-    maxInlineSize: '100%',
-    marginInline: 'auto',
-  }),
-  actionPanel: ({ euiTheme }: UseEuiTheme) =>
-    css({
-      padding: `${euiTheme.size.m} ${euiTheme.size.l}`,
-      cursor: 'pointer',
-      minWidth: 0,
-      width: '100%',
-      display: 'block',
-    }),
-  actionPanelTextWrapper: css({ minWidth: 0 }),
-};
-
-/** Applied to an action panel rendered in a disabled state (e.g. missing privileges). */
-const actionPanelDisabledStyle = css({
-  cursor: 'not-allowed',
-  opacity: 0.5,
-});
+export type ActionPolicyCreateOption = CreateOptionItem;
 
 const CREATE_WITH_AGENT_MISSING_PRIVILEGE_TOOLTIP = i18n.translate(
   'xpack.alertingV2.actionPolicyCreateOptions.createWithAgentMissingPrivilegeTooltip',
@@ -144,95 +59,32 @@ export const getCreateActionPolicyWithAgentTooltipText = ({
   return CREATE_WITH_AGENT_MISSING_SETTING_TOOLTIP;
 };
 
-const noop = () => undefined;
+interface ActionPolicyCreateOptionsPanelProps {
+  options: ActionPolicyCreateOption[];
+}
 
-const CreateOptionActionPanel: React.FC<{
-  item: ActionPolicyCreateOption;
-  actionPanelStyle: React.ComponentProps<typeof EuiPanel>['css'];
-}> = ({ item, actionPanelStyle }) => {
-  const isDisabled = item.disabled === true;
-  const hasTooltip = item.tooltipText !== undefined;
-  const panel = (
-    <EuiPanel
-      element="button"
-      hasBorder
-      paddingSize="none"
-      // EuiPanel has no built-in disabled appearance (unlike EuiCard's `isDisabled`), and a native
-      // `disabled` button would both leave it unstyled and suppress the hover events EuiToolTip
-      // needs. So we convey the disabled state with `aria-disabled` + styling and guard the click.
-      aria-disabled={isDisabled || undefined}
-      onClick={isDisabled ? noop : item.onClick}
-      css={[actionPanelStyle, isDisabled && actionPanelDisabledStyle]}
-      data-test-subj={item['data-test-subj']}
-    >
-      <EuiFlexGroup alignItems="center" gutterSize="m" responsive={false}>
-        <EuiFlexItem grow={false}>
-          <EuiIcon type={item.iconType} size="l" color="text" aria-hidden={true} />
-        </EuiFlexItem>
-        <EuiFlexItem css={listEmptyStateStyles.actionPanelTextWrapper}>
-          <EuiText size="s">
-            <strong>{item.title}</strong>
-          </EuiText>
-          <EuiText size="xs" color="subdued">
-            {item.description}
-          </EuiText>
-        </EuiFlexItem>
-      </EuiFlexGroup>
-    </EuiPanel>
-  );
-
-  if (hasTooltip) {
-    return (
-      <EuiToolTip content={item.tooltipText} display="block">
-        {panel}
-      </EuiToolTip>
-    );
-  }
-
-  return panel;
-};
-
-/** Action policies list empty state — matches the rules list create empty prompt layout. */
+/** Action policies list empty state — delegates to generic CreateOptionsPanel. */
 export const ActionPolicyCreateOptionsPanel: React.FC<ActionPolicyCreateOptionsPanelProps> = ({
   options,
-}) => {
-  const styles = useMemoCss(listEmptyStateStyles);
-
-  return (
-    <div css={styles.parent} data-test-subj="actionPolicyCreateOptionsPanel">
-      <EuiPageTemplate grow={false} offset={0} css={styles.template}>
-        <EuiPageTemplate.EmptyPrompt
-          paddingSize="none"
-          icon={
-            <EuiImage
-              size="fullWidth"
-              src={rulesListEmptyIllustration}
-              alt=""
-              data-test-subj="actionPoliciesListEmptyIllustration"
-            />
-          }
-          title={
-            <h2>
-              <FormattedMessage
-                id="xpack.alertingV2.actionPolicyCreateOptionsPanel.emptyStateTitle"
-                defaultMessage="No action policies yet. Let's get started!"
-              />
-            </h2>
-          }
-          actions={
-            <EuiFlexGroup direction="column" gutterSize="s" css={styles.actionsWrapper}>
-              {options.map((item) => (
-                <EuiFlexItem key={item.id} grow={false}>
-                  <CreateOptionActionPanel item={item} actionPanelStyle={styles.actionPanel} />
-                </EuiFlexItem>
-              ))}
-            </EuiFlexGroup>
-          }
-          titleSize="xs"
-          color="transparent"
-          css={styles.widgetContainer}
+}) => (
+  <CreateOptionsPanel
+    title={
+      <h2>
+        <FormattedMessage
+          id="xpack.alertingV2.actionPolicyCreateOptionsPanel.emptyStateTitle"
+          defaultMessage="No action policies yet. Let's get started!"
         />
-      </EuiPageTemplate>
-    </div>
-  );
-};
+      </h2>
+    }
+    icon={
+      <EuiImage
+        size="fullWidth"
+        src={rulesListEmptyIllustration}
+        alt=""
+        data-test-subj="actionPoliciesListEmptyIllustration"
+      />
+    }
+    items={options}
+    data-test-subj="actionPolicyCreateOptionsPanel"
+  />
+);
