+++ x-pack/platform/plugins/shared/alerting_v2/public/components/create_options/index.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export type { CreateOptionItem, CreateOptionsPanelProps } from './create_options_panel';
+export { CreateOptionActionPanel, CreateOptionsPanel } from './create_options_panel';
