+++ x-pack/platform/plugins/shared/alerting_v2/public/components/rule_create_options/rule_create_options_panel.tsx
@@ -13,18 +13,17 @@ import {
   EuiHorizontalRule,
   EuiIcon,
   EuiImage,
-  EuiPageTemplate,
   EuiPanel,
   EuiSpacer,
   EuiText,
   EuiTitle,
   EuiToolTip,
-  type UseEuiTheme,
 } from '@elastic/eui';
 import { css } from '@emotion/react';
-import { useMemoCss } from '@kbn/css-utils/public/use_memo_css';
 import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n-react';
+import type { CreateOptionItem } from '../create_options';
+import { CreateOptionsPanel } from '../create_options';
 import type { AgentBuilderSkillsRequirements } from '../../hooks/use_are_agent_builder_skills_available';
 import rulesListEmptyIllustration from '../../assets/illustration-results-128.svg';
 
@@ -52,67 +51,6 @@ interface RuleCreateOptionsPanelProps {
   createWithAgentTooltipText?: string;
   onCreateThresholdRule?: () => void;
   legacyRuleTypes?: LegacyRuleTypeItem[];
-}
-
-/** Fits the two primary option descriptions on one line; threshold description may wrap. */
-const LIST_EMPTY_STATE_MAX_INLINE_SIZE = '44em';
-
-const listEmptyStateStyles = {
-  parent: css({
-    display: 'flex',
-    flexGrow: 1,
-    height: '100%',
-  }),
-  template: css({
-    backgroundColor: 'inherit',
-    marginInline: 'auto',
-    maxInlineSize: LIST_EMPTY_STATE_MAX_INLINE_SIZE,
-    width: '100%',
-  }),
-  widgetContainer: ({ euiTheme }: UseEuiTheme) =>
-    css({
-      padding: euiTheme.size.xl,
-      borderRadius: euiTheme.border.radius.medium,
-      '.euiEmptyPrompt__icon': {
-        marginBottom: euiTheme.size.l,
-        paddingRight: euiTheme.size.s,
-      },
-      '.euiEmptyPrompt__content': {
-        maxInlineSize: LIST_EMPTY_STATE_MAX_INLINE_SIZE,
-        width: '100%',
-      },
-    }),
-  actionsWrapper: css({
-    width: '100%',
-    maxInlineSize: LIST_EMPTY_STATE_MAX_INLINE_SIZE,
-    marginInline: 'auto',
-  }),
-  actionPanel: ({ euiTheme }: UseEuiTheme) =>
-    css({
-      padding: `${euiTheme.size.s} ${euiTheme.size.base}`,
-      cursor: 'pointer',
-      minWidth: 0,
-    }),
-  actionPanelTextWrapper: css({ minWidth: 0 }),
-};
-
-/** Applied to an action panel rendered in a disabled state (e.g. missing privileges). */
-const actionPanelDisabledStyle = css({
-  cursor: 'not-allowed',
-  opacity: 0.5,
-});
-
-interface RuleCreateOptionItem {
-  id: string;
-  iconType: string;
-  title: string;
-  description: string;
-  onClick: () => void;
-  /** When `true`, the option is rendered disabled and its click is a no-op. */
-  disabled?: boolean;
-  /** When set, this string is shown as a hover/focus tooltip, independent of `disabled`. */
-  tooltipText?: string;
-  'data-test-subj'?: string;
 }
 
 const ESQL_RULE_TITLE = i18n.translate(
@@ -173,6 +111,7 @@ export const getCreateWithAgentTooltipText = ({
   }
   return CREATE_WITH_AGENT_MISSING_SETTING_TOOLTIP;
 };
+
 const THRESHOLD_RULE_TITLE = i18n.translate(
   'xpack.alertingV2.ruleCreateOptionsPanel.thresholdRuleTitle',
   { defaultMessage: 'Threshold rule' }
@@ -186,50 +125,87 @@ const THRESHOLD_RULE_DESCRIPTION = i18n.translate(
 
 const noop = () => undefined;
 
-const RuleCreateOptionActionPanel: React.FC<{
-  item: RuleCreateOptionItem;
-  actionPanelStyle: React.ComponentProps<typeof EuiPanel>['css'];
-}> = ({ item, actionPanelStyle }) => {
-  const isDisabled = item.disabled === true;
-  const hasTooltip = item.tooltipText !== undefined;
-  const panel = (
-    <EuiPanel
-      element="button"
-      hasBorder
-      paddingSize="none"
-      // EuiPanel has no built-in disabled appearance (unlike EuiCard's `isDisabled`), and a native
-      // `disabled` button would both leave it unstyled and suppress the hover events EuiToolTip
-      // needs. So we convey the disabled state with `aria-disabled` + styling and guard the click.
-      aria-disabled={isDisabled || undefined}
-      onClick={isDisabled ? noop : item.onClick}
-      css={[actionPanelStyle, isDisabled && actionPanelDisabledStyle]}
-      data-test-subj={item['data-test-subj']}
-    >
-      <EuiFlexGroup alignItems="center" gutterSize="m" responsive={false}>
-        <EuiFlexItem grow={false}>
-          <EuiIcon type={item.iconType} size="m" color="text" aria-hidden={true} />
-        </EuiFlexItem>
-        <EuiFlexItem css={listEmptyStateStyles.actionPanelTextWrapper}>
-          <EuiText size="s">
-            <strong>{item.title}</strong>
-          </EuiText>
-          <EuiText size="xs" color="subdued">
-            {item.description}
-          </EuiText>
-        </EuiFlexItem>
-      </EuiFlexGroup>
-    </EuiPanel>
+/** Applied to the EuiCard in the flyout layout when the option is disabled. */
+const flyoutCardDisabledStyle = css({
+  cursor: 'not-allowed',
+  opacity: 0.5,
+});
+
+/** Rules list empty state — delegates to generic CreateOptionsPanel. */
+const RuleCreateOptionsListEmptyState: React.FC<RuleCreateOptionsPanelProps> = ({
+  onCreateEsqlRule,
+  onCreateWithAgent,
+  createWithAgentDisabled,
+  createWithAgentTooltipText,
+  onCreateThresholdRule,
+}) => {
+  const primaryItems = useMemo<CreateOptionItem[]>(
+    () => [
+      {
+        id: 'create-esql-rule',
+        iconType: 'productDiscover',
+        title: ESQL_RULE_TITLE,
+        description: ESQL_RULE_DESCRIPTION,
+        onClick: onCreateEsqlRule,
+        'data-test-subj': 'createEsqlRuleCard',
+      },
+      {
+        id: 'create-with-agent',
+        iconType: 'productAgent',
+        title: AI_AGENT_TITLE,
+        description: AI_AGENT_DESCRIPTION,
+        onClick: onCreateWithAgent,
+        disabled: createWithAgentDisabled,
+        tooltipText: createWithAgentTooltipText,
+        'data-test-subj': 'createWithAgentCard',
+      },
+    ],
+    [onCreateEsqlRule, onCreateWithAgent, createWithAgentDisabled, createWithAgentTooltipText]
   );
 
-  if (hasTooltip) {
-    return (
-      <EuiToolTip content={item.tooltipText} display="block">
-        {panel}
-      </EuiToolTip>
-    );
-  }
+  const secondaryItems = useMemo<CreateOptionItem[]>(
+    () => [
+      {
+        id: 'create-threshold-rule',
+        iconType: 'chartThreshold',
+        title: THRESHOLD_RULE_TITLE,
+        description: THRESHOLD_RULE_DESCRIPTION,
+        onClick: onCreateThresholdRule ?? noop,
+        'data-test-subj': 'createThresholdRuleCard',
+      },
+    ],
+    [onCreateThresholdRule]
+  );
 
-  return panel;
+  return (
+    <CreateOptionsPanel
+      title={
+        <h2>
+          <FormattedMessage
+            id="xpack.alertingV2.ruleCreateOptionsPanel.emptyStateTitle"
+            defaultMessage="No rules yet. Let's get started!"
+          />
+        </h2>
+      }
+      icon={
+        <EuiImage
+          size="fullWidth"
+          src={rulesListEmptyIllustration}
+          alt=""
+          data-test-subj="rulesListEmptyIllustration"
+        />
+      }
+      items={primaryItems}
+      secondaryItems={secondaryItems}
+      secondaryLabel={
+        <FormattedMessage
+          id="xpack.alertingV2.ruleCreateOptionsPanel.orStartFromBuilderLabel"
+          defaultMessage="or start from a builder"
+        />
+      }
+      data-test-subj="ruleCreateOptionsPanel"
+    />
+  );
 };
 
 const RuleBuilderSectionDivider: React.FC = () => (
@@ -254,98 +230,6 @@ const RuleBuilderSectionDivider: React.FC = () => (
     <EuiSpacer size="l" />
   </>
 );
-
-/** Rules list empty state — matches dashboard create empty prompt layout. */
-const RuleCreateOptionsListEmptyState: React.FC<RuleCreateOptionsPanelProps> = ({
-  onCreateEsqlRule,
-  onCreateWithAgent,
-  createWithAgentDisabled,
-  createWithAgentTooltipText,
-  onCreateThresholdRule,
-}) => {
-  const styles = useMemoCss(listEmptyStateStyles);
-
-  const primaryCreateOptions = useMemo<RuleCreateOptionItem[]>(
-    () => [
-      {
-        id: 'create-esql-rule',
-        iconType: 'productDiscover',
-        title: ESQL_RULE_TITLE,
-        description: ESQL_RULE_DESCRIPTION,
-        onClick: onCreateEsqlRule,
-        'data-test-subj': 'createEsqlRuleCard',
-      },
-      {
-        id: 'create-with-agent',
-        iconType: 'productAgent',
-        title: AI_AGENT_TITLE,
-        description: AI_AGENT_DESCRIPTION,
-        onClick: onCreateWithAgent,
-        disabled: createWithAgentDisabled,
-        tooltipText: createWithAgentTooltipText,
-        'data-test-subj': 'createWithAgentCard',
-      },
-    ],
-    [onCreateEsqlRule, onCreateWithAgent, createWithAgentDisabled, createWithAgentTooltipText]
-  );
-
-  const thresholdCreateOption = useMemo<RuleCreateOptionItem>(
-    () => ({
-      id: 'create-threshold-rule',
-      iconType: 'chartThreshold',
-      title: THRESHOLD_RULE_TITLE,
-      description: THRESHOLD_RULE_DESCRIPTION,
-      onClick: onCreateThresholdRule ?? noop,
-      'data-test-subj': 'createThresholdRuleCard',
-    }),
-    [onCreateThresholdRule]
-  );
-
-  return (
-    <div css={listEmptyStateStyles.parent} data-test-subj="ruleCreateOptionsPanel">
-      <EuiPageTemplate grow={false} offset={0} css={styles.template}>
-        <EuiPageTemplate.EmptyPrompt
-          paddingSize="none"
-          icon={
-            <EuiImage
-              size="fullWidth"
-              src={rulesListEmptyIllustration}
-              alt=""
-              data-test-subj="rulesListEmptyIllustration"
-            />
-          }
-          title={
-            <h2>
-              <FormattedMessage
-                id="xpack.alertingV2.ruleCreateOptionsPanel.emptyStateTitle"
-                defaultMessage="No rules yet. Let's get started!"
-              />
-            </h2>
-          }
-          actions={
-            <EuiFlexGroup direction="column" gutterSize="s" css={styles.actionsWrapper}>
-              {primaryCreateOptions.map((item) => (
-                <EuiFlexItem key={item.id} grow={false}>
-                  <RuleCreateOptionActionPanel item={item} actionPanelStyle={styles.actionPanel} />
-                </EuiFlexItem>
-              ))}
-              <EuiFlexItem grow={false}>
-                <RuleBuilderSectionDivider />
-                <RuleCreateOptionActionPanel
-                  item={thresholdCreateOption}
-                  actionPanelStyle={styles.actionPanel}
-                />
-              </EuiFlexItem>
-            </EuiFlexGroup>
-          }
-          titleSize="xs"
-          color="transparent"
-          css={styles.widgetContainer}
-        />
-      </EuiPageTemplate>
-    </div>
-  );
-};
 
 const LegacyRuleTypesSection: React.FC<{ items: LegacyRuleTypeItem[] }> = ({ items }) => {
   if (items.length === 0) return null;
@@ -408,13 +292,8 @@ const RuleCreateOptionsFlyoutPanel: React.FC<RuleCreateOptionsPanelProps> = ({
       titleElement="h3"
       titleSize="xs"
       hasBorder={true}
-      // Convey the disabled state with `aria-disabled` + styling and a guarded click rather than
-      // EuiCard's native `isDisabled` (which renders a `disabled` <button>). A native disabled
-      // control drops out of the tab order and suppresses the hover events EuiToolTip needs, so the
-      // explanatory tooltip would be unreachable — defeating the goal of signalling how to gain
-      // access. Keeping it focusable lets the tooltip surface on both hover and keyboard focus.
       aria-disabled={isAgentDisabled || undefined}
-      css={isAgentDisabled ? actionPanelDisabledStyle : undefined}
+      css={isAgentDisabled ? flyoutCardDisabledStyle : undefined}
       title={AI_AGENT_TITLE}
       description={AI_AGENT_DESCRIPTION}
       onClick={isAgentDisabled ? noop : onCreateWithAgent}
